Move PolygonInteriorIterator
