+++ src/scalar/polygon/mod.rs
@@ -1,5 +1,7 @@
+pub mod iterator;
 mod owned;
 mod scalar;
 
+pub use iterator::PolygonInteriorIterator;
 pub use owned::OwnedPolygon;
 pub use scalar::Polygon;
